Cast the pointer token for the thread ID to a guint to remove the warning.

+++ logger/eas-logger.c
@@ -37,32 +37,32 @@ void eas_logger (const gchar *log_domain,
     if (logfile)
     {
         if (envLevel > 0 && log_level == G_LOG_LEVEL_CRITICAL)
-            fprintf (logfile, "(process:%d:%x): %s%s*** CRITICAL ***:%s \n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (logfile, "(process:%d:%x): %s%s*** CRITICAL ***:%s \n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 1 && log_level == G_LOG_LEVEL_WARNING)
-            fprintf (logfile, "(process:%d:%x): %s%sWARNING **:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (logfile, "(process:%d:%x): %s%sWARNING **:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 2 && log_level == G_LOG_LEVEL_MESSAGE)
-            fprintf (logfile, "(process:%d:%x): %s%sMESSAGE:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (logfile, "(process:%d:%x): %s%sMESSAGE:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 3 && log_level == G_LOG_LEVEL_DEBUG)
-            fprintf (logfile, "(process:%d:%x): %s%sDEBUG:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (logfile, "(process:%d:%x): %s%sDEBUG:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         fclose (logfile);
     }
     else
     {
         if (envLevel > 0 && log_level == G_LOG_LEVEL_CRITICAL)
-            fprintf (stderr, "(process:%d:%x): %s%s*** CRITICAL ***:%s \n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (stderr, "(process:%d:%x): %s%s*** CRITICAL ***:%s \n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 1 && log_level == G_LOG_LEVEL_WARNING)
-            fprintf (stderr, "(process:%d:%x): %s%sWARNING **:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (stderr, "(process:%d:%x): %s%sWARNING **:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 2 && log_level == G_LOG_LEVEL_MESSAGE)
-            fprintf (stderr, "(process:%d:%x): %s%sMessage:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (stderr, "(process:%d:%x): %s%sMessage:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
 
         if (envLevel > 3 && log_level == G_LOG_LEVEL_DEBUG)
-            fprintf (stdout, "(process:%d:%x): %s%sDEBUG:%s\n", pid, tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
+            fprintf (stdout, "(process:%d:%x): %s%sDEBUG:%s\n", pid, (guint)tid, (log_domain ?: ""), (log_domain ? "-" : ""), message);
             
         fflush(stderr);
         fflush (stdout);
